:recycle: api: Move locals constants to src
