+++ api/lib/domain/constants.js
@@ -51,17 +51,6 @@ const PIX_CERTIF = {
   CURRENT_CERTIFICATION_VERSION: 2,
 };
 
-const LOCALE = {
-  ENGLISH_SPOKEN: 'en',
-  FRENCH_FRANCE: 'fr-fr',
-  FRENCH_SPOKEN: 'fr',
-};
-const LANG = {
-  ENGLISH: 'en',
-  FRENCH: 'fr',
-};
-const SUPPORTED_LOCALES = ['en', 'fr', 'fr-BE', 'fr-FR'];
-
 const STUDENT_RECONCILIATION_ERRORS = {
   RECONCILIATION: {
     IN_OTHER_ORGANIZATION: {
@@ -143,9 +132,7 @@ const constants = {
   PIX_ORGA,
   PIX_ADMIN,
   PIX_CERTIF,
-  LOCALE,
   STUDENT_RECONCILIATION_ERRORS,
-  SUPPORTED_LOCALES,
   OIDC_ERRORS,
   CERTIFICATION_CENTER_TYPES,
   ORGANIZATION_FEATURE,
@@ -177,10 +164,7 @@ export {
   PIX_ORGA,
   PIX_ADMIN,
   PIX_CERTIF,
-  LANG,
-  LOCALE,
   STUDENT_RECONCILIATION_ERRORS,
-  SUPPORTED_LOCALES,
   OIDC_ERRORS,
   CERTIFICATION_CENTER_TYPES,
   ORGANIZATION_FEATURE,
