+++ api/lib/infrastructure/repositories/thematic-repository.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import { Thematic } from '../../domain/models/Thematic.js';
 import { thematicDatasource } from '../datasources/learning-content/thematic-datasource.js';
 import { getTranslatedKey } from '../../domain/services/get-translated-text.js';
-import { LOCALE } from '../../domain/constants.js';
+import { LOCALE } from '../../../src/shared/domain/constants.js';
 
 const { FRENCH_FRANCE } = LOCALE;
