+++ api/lib/infrastructure/utils/request-response-utils.js
@@ -1,6 +1,6 @@
 import accept from '@hapi/accept';
 import { tokenService } from '../../domain/services/token-service.js';
-import { LOCALE } from '../../domain/constants.js';
+import { LOCALE } from '../../../src/shared/domain/constants.js';
 
 const { ENGLISH_SPOKEN, FRENCH_FRANCE, FRENCH_SPOKEN } = LOCALE;
