+++ api/scripts/certification/generate-certification-test-statistics.js
@@ -12,7 +12,7 @@ import { knex, disconnect } from '../../db/knex-database-connection.js';
 import * as competenceRepository from '../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as placementProfileService from '../../lib/domain/services/placement-profile-service.js';
 import * as certificationChallengeService from '../../lib/domain/services/certification-challenges-service.js';
-import { LOCALE } from '../../lib/domain/constants.js';
+import { LOCALE } from '../../src/shared/domain/constants.js';
 import * as url from 'url';
 
 const { FRENCH_FRANCE } = LOCALE;
