+++ api/src/shared/infrastructure/repositories/competence-repository.js
@@ -3,7 +3,8 @@ import { LearningContentResourceNotFound } from '../../../../lib/infrastructure/
 import { Competence } from '../../domain/models/Competence.js';
 import { competenceDatasource } from '../../../../lib/infrastructure/datasources/learning-content/competence-datasource.js';
 import { NotFoundError } from '../../../../lib/domain/errors.js';
-import { LOCALE, PIX_ORIGIN } from '../../../../lib/domain/constants.js';
+import { PIX_ORIGIN } from '../../../../lib/domain/constants.js';
+import { LOCALE } from '../../../../src/shared/domain/constants.js';
 
 const { FRENCH_FRANCE } = LOCALE;
