+++ api/tests/acceptance/application/scorecards/scorecard-controller_test.js
@@ -7,7 +7,7 @@ import {
 } from '../../../test-helper.js';
 
 import { KnowledgeElement } from '../../../../lib/domain/models/KnowledgeElement.js';
-import { LOCALE } from '../../../../lib/domain/constants.js';
+import { LOCALE } from '../../../../src/shared/domain/constants.js';
 
 const { FRENCH_SPOKEN } = LOCALE;
