+++ api/tests/integration/domain/services/placement-profile-service_test.js
@@ -1,7 +1,7 @@
 import { expect, databaseBuilder, mockLearningContent, learningContentBuilder } from '../../../test-helper.js';
 import * as placementProfileService from '../../../../lib/domain/services/placement-profile-service.js';
 import { KnowledgeElement } from '../../../../lib/domain/models/KnowledgeElement.js';
-import { LOCALE } from '../../../../lib/domain/constants.js';
+import { LOCALE } from '../../../../src/shared/domain/constants.js';
 
 const { ENGLISH_SPOKEN } = LOCALE;
