+++ api/tests/integration/domain/usecases/get-campaign-profile_test.js
@@ -1,7 +1,6 @@
 import { expect, databaseBuilder, mockLearningContent } from '../../../test-helper.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
-import { LOCALE } from '../../../../lib/domain/constants.js';
-
+import { LOCALE } from '../../../../src/shared/domain/constants.js';
 const { FRENCH_SPOKEN } = LOCALE;
 
 describe('Integration | UseCase | get-campaign-profile', function () {
