+++ api/tests/integration/infrastructure/repositories/campaign-profile-repository_test.js
@@ -1,6 +1,8 @@
 import { expect, databaseBuilder, mockLearningContent, catchErr } from '../../../test-helper.js';
 import * as CampaignProfileRepository from '../../../../lib/infrastructure/repositories/campaign-profile-repository.js';
-import { PIX_COUNT_BY_LEVEL, LOCALE } from '../../../../lib/domain/constants.js';
+import { PIX_COUNT_BY_LEVEL } from '../../../../lib/domain/constants.js';
+import { LOCALE } from '../../../../src/shared/domain/constants.js';
+
 import { NotFoundError } from '../../../../lib/domain/errors.js';
 
 const { ENGLISH_SPOKEN, FRENCH_SPOKEN } = LOCALE;
