+++ api/tests/unit/application/assessments/assessment-controller-get-next-challenge_test.js
@@ -2,7 +2,7 @@ import { sinon, expect, domainBuilder, generateValidRequestAuthorizationHeader }
 import { assessmentController } from '../../../../lib/application/assessments/assessment-controller.js';
 import { AssessmentEndedError } from '../../../../lib/domain/errors.js';
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
-import { LOCALE } from '../../../../lib/domain/constants.js';
+import { LOCALE } from '../../../../src/shared/domain/constants.js';
 
 const { FRENCH_FRANCE, FRENCH_SPOKEN } = LOCALE;
