+++ api/tests/unit/application/campaign/campaign-controller_test.js
@@ -2,7 +2,7 @@ import { sinon, expect, domainBuilder, hFake, catchErr } from '../../../test-hel
 import { campaignController } from '../../../../lib/application/campaigns/campaign-controller.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
 import { UserNotAuthorizedToAccessEntityError } from '../../../../lib/domain/errors.js';
-import { LOCALE } from '../../../../lib/domain/constants.js';
+import { LOCALE } from '../../../../src/shared/domain/constants.js';
 import { ForbiddenAccess } from '../../../../src/shared/domain/errors.js';
 
 const { FRENCH_SPOKEN } = LOCALE;
