+++ api/tests/unit/domain/services/mail-service_test.js
@@ -13,7 +13,7 @@ const mainTranslationsMapping = {
   en,
 };
 
-import { LOCALE } from '../../../../lib/domain/constants.js';
+import { LOCALE } from '../../../../src/shared/domain/constants.js';
 
 const { ENGLISH_SPOKEN, FRENCH_FRANCE, FRENCH_SPOKEN } = LOCALE;
