+++ api/tests/unit/domain/usecases/compute-campaign-analysis_test.js
@@ -1,7 +1,7 @@
 import { expect, sinon, catchErr, domainBuilder } from '../../../test-helper.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
 import { UserNotAuthorizedToAccessEntityError } from '../../../../lib/domain/errors.js';
-import { LOCALE } from '../../../../lib/domain/constants.js';
+import { LOCALE } from '../../../../src/shared/domain/constants.js';
 
 const { FRENCH_SPOKEN } = LOCALE;
 const { computeCampaignAnalysis } = usecases;
